CoxeterGroups.jl Documentation

This package is for creating and working with Coxeter groups.

Creating a group

There are various implementations of Coxeter groups, each represented by a concrete type which is a subtype of the abstract CoxGrp. Use one of the constructors below to create a group, depending on which implementation is desired. For example:

# A Coxeter group of type A2, implemented using minimal roots.
W, (s, t) = coxter_group_min([2, -1; -1, 2])

# The symmetric group S3, implemented via permutations.
S, (u, v) = symmetric_group(3)

Here is a summary of the creation functions - there is more information on specific implementations in the Implementations section.

CoxeterGroups.coxeter_group_min — Function

coxeter_group_min(mat) -> CoxGrpMin, gens

Given a Coxeter matrix or GCM, create a Coxeter group implemented via the minimal roots automaton. Currently only a GCM can be passed.

CoxeterGroups.symmetric_group — Function

symmetric_group(n) -> CoxGrpSym, gens

Create the symmetric group on $n$ elements (a Coxeter system of type $A_{n-1}$), for $n \geq 0$.

CoxeterGroups.coxeter_group_recursive — Function

coxeter_group_recursive(M, [names])

Create a Coxeter group with elements represented in InverseShortLex normal form, with group operations implemented with a recursive algorithm.

coxeter_group_recursive(groupType::String)

Create a Coxeter Group based on one or more group type strings. To denote the direct product of group types, enter the group types seperated by the 'x' character. Affine groups are denoted by including a '~' character after a letter.

#Examples

julia> CG, _ = coxeter_group_recursive("A3")
(CoxGrpRec([1 3 2; 3 1 3; 2 3 1], ["<1>", "<2>", "<3>"]), CoxEltRec[CoxEltRec(CoxGrpRec([1 3 2; 3 1 3; 2 3 1], ["<1>", "<2>", "<3>"]), [1]), CoxEltRec(CoxGrpRec([1 3 2; 3 1 3; 2 3 1], ["<1>", "<2>", "<3>"]), [2]), CoxEltRec(CoxGrpRec([1 3 2; 3 1 3; 2 3 1], ["<1>", "<2>", "<3>"]), [3])])

julia> CG
Coxeter Group with Coxeter Matrix:
3×3 Matrix{Int64}:
 1  3  2
 3  1  3
 2  3  1
and generating Set:
["<1>", "<2>", "<3>"]

julia> CG, _ = coxeter_group_recursive("A3 x B4")
[1 3 2; 3 1 3; 2 3 1]
[1 3 2 2; 3 1 3 2; 2 3 1 4; 2 2 4 1]
(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), CoxEltRec[CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [1]), CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [2]), CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [3]), CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [4]), CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [5]), CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [6]), CoxEltRec(CoxGrpRec([1 3 … 2 2; 3 1 … 2 2; … ; 2 2 … 1 4; 2 2 … 4 1], ["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]), [7])])

julia> CG
Coxeter Group with Coxeter Matrix:
7×7 Matrix{Int64}:
 1  3  2  2  2  2  2
 3  1  3  2  2  2  2
 2  3  1  2  2  2  2
 2  2  2  1  3  2  2
 2  2  2  3  1  3  2
 2  2  2  2  3  1  4
 2  2  2  2  2  4  1
and generating Set:
["<1>", "<2>", "<3>", "<4>", "<5>", "<6>", "<7>"]

julia> CG, _ = coxeter_group_recursive("D~4")
(CoxGrpRec([1 2 … 2 2; 2 1 … 2 2; … ; 2 2 … 1 2; 2 2 … 2 1], ["<1>", "<2>", "<3>", "<4>", "<5>"]), CoxEltRec[CoxEltRec(CoxGrpRec([1 2 … 2 2; 2 1 … 2 2; … ; 2 2 … 1 2; 2 2 … 2 1], ["<1>", "<2>", "<3>", "<4>", "<5>"]), [1]), CoxEltRec(CoxGrpRec([1 2 … 2 2; 2 1 … 2 2; … ; 2 2 … 1 2; 2 2 … 2 1], ["<1>", "<2>", "<3>", "<4>", "<5>"]), [2]), CoxEltRec(CoxGrpRec([1 2 … 2 2; 2 1 … 2 2; … ; 2 2 … 1 2; 2 2 … 2 1], ["<1>", "<2>", "<3>", "<4>", "<5>"]), [3]), CoxEltRec(CoxGrpRec([1 2 … 2 2; 2 1 … 2 2; … ; 2 2 … 1 2; 2 2 … 2 1], ["<1>", "<2>", "<3>", "<4>", "<5>"]), [4]), CoxEltRec(CoxGrpRec([1 2 … 2 2; 2 1 … 2 2; … ; 2 2 … 1 2; 2 2 … 2 1], ["<1>", "<2>", "<3>", "<4>", "<5>"]), [5])])

julia> CG
Coxeter Group with Coxeter Matrix:
5×5 Matrix{Int64}:
 1  2  3  2  2
 2  1  3  2  2
 3  3  1  3  3
 2  2  3  1  2
 2  2  3  2  1
and generating Set:
["<1>", "<2>", "<3>", "<4>", "<5>"]

Operations on the group

A Coxeter group is a subtype of the abstract CoxGrp type. Coxeter groups compare equal by reference. The following operations are supported on all group implementations:

CoxeterGroups.rank — Method

rank(W::CoxGrp) -> Integer

Return the rank (number of generators) of the Coxeter group.

CoxeterGroups.generators — Method

generators(W::CoxGrp)

Return the simple generators of $W$.

Base.one — Method

one(W::CoxGrp) -> CoxElt

Return the identity in the Coxeter group $W$.

CoxeterGroups.coxeter_matrix — Method

coxeter_matrix(W::CoxGrp) -> Matrix

Return the Coxeter matrix for this group.

CoxeterGroups.is_finite — Method

is_finite(W::CoxGrp) -> Bool

Return whether the group $W$ is finite or not.

CoxeterGroups.longest_element — Method

longest_element(W::CoxGrp)

Return the longest element if $W$ is finite, otherwise error.

Operations on group elements

A Coxeter group element is a subtype of the abstract CoxElt type. Coxeter group elements are immutable value types: they compare equal if they represent the same group element, they can be used as keys in hash tables, and so on. The following operations are supported on all Coxeter group element implementations:

Base.parent — Method

parent(w::CoxElt) -> CoxGrp

Return the parent group of $W$.

Base.isone — Method

isone(w::CoxElt) -> Bool

Check if $w$ is the identity of the group.

Base.:* — Method

*(w, x)

The product $wx$ of Coxeter group elements.

Base.inv — Method

inv(w)

The inverse $w^{-1}$ of a Coxeter group element.

Base.:^ — Method

^(x::CoxElt, e::T) where T <: Integer

Return the power $x^e$.

Base.length — Method

length(w::CoxElt) -> Integer

The length of the group element $w$, i.e. the number of generators in any reduced expression for $w$.

Base.sign — Method

sign(w::CoxElt) -> Int

The sign homomorphism $(-1)^{l(w)}$.

CoxeterGroups.is_right_descent — Method

is_right_descent(w::CoxElt, t::Integer) -> Bool

Return true if $t$ is a right descent of $w$, i.e. if $l(wt) < l(w)$.

CoxeterGroups.is_left_descent — Method

is_left_descent(t::Integer, w::CoxElt) -> Bool

Return true if $t$ is a left descent of $w$, i.e. if $l(tw) < l(w)$.

CoxeterGroups.right_multiply — Method

right_multiply(w::CoxElt, s::Integer) -> CoxElt

Return the product $ws$, where $s$ is the simple reflection indexed by $s$.

CoxeterGroups.left_multiply — Method

left_multiply(s::Integer, w::CoxElt) -> CoxElt

Return the product $sw$, where $s$ is the simple reflection indexed by $s$.

CoxeterGroups.short_lex — Method

short_lex(w::CoxElt) -> Vector{Int8}

Return the ShortLex normal form of $w$, meaning the lexicographically least reduced expression.

CoxeterGroups.inverse_short_lex — Method

inverse_short_lex(w)

Return the InverseShortLex normal form of w: the lexicographically least reduced expression for $w$ when read backwards, or equivalently the reverse of the ShortLex normal form of $w^{-1}$.

Internals

A Coxeter group implementation consists of a new pair of concrete types subtyping CoxGrp and CoxElt respectively.

CoxeterGroups.CoxGrp — Type

The abstract supertype for Coxeter groups.

CoxeterGroups.CoxElt — Type

The abstract supertype for Coxeter group elements.

When adding a new concrete implementation, a small set of functions must be implemented explicitly, then generic implementations of other functions will begin working. The concrete implementation should specialise these generic implementations where it is possible to take advantage of specific internal structure.

The group type must implement a constructor such as those in Creating a group, and the functions coxeter_matrix rank, one, generators, and is_finite.

The element type must implement parent, ==, hash, is_left_descent, is_right_descent, left_multiply, and right_multiply.

The following functions will then be defined:

For the group type: longest_element.

For the element type: isone, length, sign, short_lex, inverse_short_lex, *, inv, ^.

Code conventions

Code conventions are (currently) as follows:

Library users should access data via methods, not via properties.

A Coxeter group implementation called foo should define a concrete subtype CoxGrpFoo of CoxGrp, and a concrete subtype CoxEltFoo of CoxElt.

Mathematical conventions

A generalised Cartan matrix or GCM is a square matrix $A \in \operatorname{Mat}_I(\mathbb{Z})$ over the integers which satisfies the following properties:

(C1) All diagonal elements are equal to $2$,

(C2) All off-diagonal entries are zero or negative, and

(C3) $A_{ij} = 0$ if and only if $A_{ji} = 0$.

A Coxeter-Cartan matrix is a square matrix $A$ over the reals which satisfies the conditions C1, C2, C3 of a GCM, and additionally

(C4) The product $A_{ij} A_{ji}$ is either equal to $4 \cos^2(\pi/m)$ for some $m \in \{2, 3, \ldots\}$, or $A_{ij} A_{ji} \geq 4$.

A based root datum over the PID $R$ (we will only be thinking of $R = \mathbb{Z}$ or $R = \mathbb{R}$) is the data of:

Two free $R$-modules $X$ and $X^\vee$,

A perfect pairing $\langle -, - \rangle \colon X^\vee \times X \to R$,

A collection of simple roots $\Delta = \{\alpha_i\}_{i \in I} \in X$,

A collection of simple coroots $\Delta^\vee = \{\alpha_i^\vee\}_{i \in I} \in X^\vee$,

such that the matrix $A_{ij} := \langle \alpha_i^\vee, \alpha_j \rangle$ is a GCM. We then say that the based root datum is a realisation of the GCM $A$. Define the reflections $r_i \colon X \to X$ and $r_i^\vee \colon X^\vee \to X^\vee$ by the formulas

\[r_i(\lambda) = \lambda - \langle \alpha_i^\vee, \lambda \rangle \alpha_i, \quad r_i^\vee(\nu) = \nu - \langle \nu, \alpha_i \rangle \alpha_i^\vee.\]

Then the reflections each define a faithful representation of a Coxeter group $(W, I)$ on each of $X$ and $X^\vee$, which are dual in the sense that $\langle w \nu, \lambda \rangle = \langle \nu, w \lambda \rangle$.

Every GCM $A$ admits two canonical based root data. The simply-connected realisation is the root datum where $X^\vee$ is the free $R$-module with basis $(\alpha_i^\vee)_{i \in I}$, $X = \operatorname{Hom}(X^\vee, R)$ is the dual, the pairing $\langle -, - \rangle$ is the canonical pairing between a module and its dual, and the simple roots $\alpha_i$ are uniquely determined by the condition that $\langle \alpha_i^\vee, \alpha_j \rangle = A_{ij}$. The adjoint realisation is the dual of the simply-connected realisation.

In the reflection representation implementation, we use the adjoint realisation of a GCM. Concretely, the space $X$ is the space of column vectors, with each simple root $\alpha_i$ a coordinate vector, the space $X^\vee$ is the space of row vectors, with each simple coroot $\alpha_i^\vee$ equal to the $i$th row of the Cartan matrix, and the pairing $\langle -, - \rangle$ is the dot product.

Implementations

Minimal roots implementation

The minimal roots implementation of a Coxeter group represents group elements by words in the Coxeter generators in ShortLex normal form. Operations on these words are done with the assistance of a minimal root reflection table, which needs to be constructed when the group is created.

Currently this implementation can handle any Coxeter group defined by a generalised Cartan matrix, but this is only because generating the minimal root reflection table is more difficult for general Coxeter groups. The same underlying algorithms will work once given a new table.

Performance characteristics for a Coxeter group of rank $r$:

The group itself takes $O(r M)$ space, where $M$ is the number of minimal roots. For finite types, the minimal roots are exactly the positive roots, and for affine types, there are twice as many minimal roots as there are positive roots for the corresponding finite system. The number of roots in a finite system tends to be $O(r^2)$ in the rank $r$.

Length is an $O(1)$ operation.

ShortLex normal form should be $O(1)$, but is $O(l(w))$ since it returns a defensive copy of the internal word.

Testing a left or right descent is $O(l(w))$.

Multiplication on the right $xs$ by a simple generator $s$ is $O(l(x))$.

Multiplication on the left $sx$ by a simple generator $s$ is $O(l(x)^2)$.

Multiplication $xy$ is $O(l(y)(l(x) + l(y)))$.

Inversion $x^{-1}$ is $O(l(x)^2)$.

CoxeterGroups.CoxGrpMin — Type

A Coxeter group, implemented using the minimal root reflection table.

CoxeterGroups.create_reflection_table_gcm — Function

create_reflection_table_gcm(gcm)

Create the minimal root reflection table for a Coxeter group defined by a GCM (generalised Cartan matrix). This is a straightforward algorithm since we need only deal with integers, rather than algebraic integers.

The return value is a matrix refl_table of size R×T, where R is the rank of the group (the size of the square GCM), and T is the number of minimal roots. The minimal roots are indexed from 1:T, with the first 1:R of them corresponding to the simple roots, and the other indices arbitrary.

If $β = s(α)$ is a minimal root, then refl_table[s, α] stores the index of β, and otherwise refl_table[s, α] = 0. Note that refl_table[s, s] = 0 for every simple root s.

Symmetric groups

The symmetric group $S_n$ on $n$ letters is the group of permutations of the set $[n] = \{1, \ldots, n\}$, whose standard Coxeter generators are $s_1, \ldots, s_{n-1}$ where $s_i = (i, i+1)$ is an adjacent transposition. The symmetric group may be constructed using the symmetric_group function.

A permutation $\pi \colon [n] \to [n]$ is represented as a pair of arrays recording both $\pi$ and its inverse $\pi^{-1}$:

\[ \pi \mapsto (\mathtt{fwd} = [\pi(1), \ldots, \pi(n)], \mathtt{inv} = [\pi^{-1}(1), \ldots, \pi^{-1}(n)]).\]

Complexity of operations in terms of $n$:

Creation of the group is $O(1)$.

Memory usage for each element is $O(n)$.

Testing a left or right descent is $O(1)$.

Multiplication is $O(n)$.

Calculating the sign of a permutation is $O(n)$

Taking Coxeter length takes $O(n^2)$ time (this could be improved to $O(n \log n)$ by a standard divide-and-conquer trick to count inversions).

Creating Coxeter systems

A Coxeter system is an object which stores a Coxeter matrix along with a classification into components of known type. Currently all finite and affine type components are classified, but anything indefinite (hyperbolic or otherwise) will be classified as Unknown type. Some properties of finite-type Coxeter systems can be inferred entirely from this classification, without having to construct a group object at all. For instance:

using CoxeterGroups
gcm = [
     2  -1   0   0
    -1   2  -1   0
     0  -2   2   0
     0   0   0   2
]
cox = coxeter_system(gcm)
cox

Coxeter system (rank 4, reducible, finite type) of type C3 x A1

Data known purely by classification can then be queried:

rank(cox), coxeter_name(cox), number_of_reflections(cox), order(cox)

(4, "C3 x A1", 10, 96)

The construction methods and information methods that work with Coxeter systems is below:

CoxeterGroups.coxeter_system — Function

coxeter_system(mat)

Create a Coxeter system from a Coxeter matrix or Cartan matrix.

CoxeterGroups.coxeter_name — Method

coxeter_name(cox::CoxeterSystem)

A string representing the type of the system, such as "H3 x A2".

CoxeterGroups.rank — Method

rank(cox::CoxeterSystem)

The rank of a Coxeter system $(W, S)$ is the number $|S|$ of simple generators.

CoxeterGroups.is_irreducible — Method

is_irreducible(cox::CoxeterSystem)

A Coxeter system $(W, S)$ is irreducible if its associated underyling graph consists of a single connected component. In particular, $|S| ≥ 1$.

CoxeterGroups.is_finite_type — Method

is_finite_type(cox::CoxeterSystem)

A Coxeter system $(W, S)$ is finite type if the group $W$ is finite.

CoxeterGroups.is_affine_type — Method

is_affine_type(cox::CoxeterSystem)

A Coxeter system $(W, S)$ is affine type if is irreducible, and the unique component is of affine type.

CoxeterGroups.order — Method

order(cox::CoxeterSystem)

The order of a Coxeter system $(W, S)$ is the order of the group $W$. Throws an error if the system is not finite type.

CoxeterGroups.coxeter_number — Method

coxeter_number(cox::CoxeterSystem)

The Coxeter number $h$ of an irreducible Coxeter system $(W, S)$ is the order of any Coxeter element (product of all simple reflections), or $2 |Φ^+| / |S|$, or the largest degree $d_{|S|}$. If the Coxeter system is not irreducible then these definitions all diverge, so this function will throw an error on a reducible system.

Throws an error if the system is reducible or not of finite type.

CoxeterGroups.number_of_reflections — Method

number_of_reflections(cox::CoxeterSystem)

The number of reflections in a Coxeter system $(W, S)$ is the number of conjugates of the simple generators $S$, or equivalently the number of roots in a root system for $W$. Throws an error if the system is not finite type.

CoxeterGroups.degrees — Method

degrees(cox::CoxeterSystem)

The degrees $d_1 ≤ ⋯ ≤ d_r$ of a Coxeter system $(W, S)$ are the degrees of the fundamental invariants for the action of $W$ on $\operatorname{Sym}(V^*)$, where $V$ is an irreducible representation of $W$ with $S$ acting by reflections. Throws an error if the system is not finite type.

CoxeterGroups.exponents — Method

exponents(cox::CoxeterSystem)

The exponents $m_1 ≤ ⋯ ≤ m_r$ of a Coxeter system $(W, S)$ describe the eigenvalues (with multiplicity) of a Coxeter element acting on the Tits representation $V$, which are roots of unity $\exp(2 \pi i m_i / h)$ where $h$ is the Coxeter number. Throws an error if the system is not finite type.